// src/equipos-t/dto/create-equipos-t.dto.d.ts
export declare class CreateEquiposTDto {
    nombre: string;
    cel_guardia: string;
    codigo: string;
    codigo_postal: string;
    direccion: string;
    email: string;
    fax: string;
    observacion: string;
    responsable: string;
    telefono: string;
    usuario_alta_id: number;
    usuario_modif_id: number;
    provincia_id: number;
}